coroipcc.h (coroipcc_msg_send_reply_receive): Make res_len size_t.
* include/corosync/coroipcc.h (coroipcc_msg_send_reply_receive):
change type of res_len parameter.
* lib/coroipcc.c (coroipcc_reply_receive): Likewise.

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2045 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/corosync/coroipcc.h
@@ -98,7 +98,7 @@ coroipcc_msg_send_reply_receive (
 	const struct iovec *iov,
 	unsigned int iov_len,
 	void *res_msg,
-	int res_len);
+	size_t res_len);
 
 cs_error_t
 coroipcc_msg_send_reply_receive_in_buf (
